+++ kernel/arch/x86_64/paging/x86_temp_mapper.cc
@@ -43,7 +43,7 @@ x86_PAE_Temp_Mapper::x86_PAE_Temp_Mapper(void *virt_addr, u64 cr3)
 
     Page_Table_Argumments arg {1, 1, 0, 0, 1, 000};
 
-    auto pt_phys = prepare_pt_for((u64)virt_addr, arg, cr3);
+    auto pt_phys = prepare_pt_for(virt_addr, arg, cr3);
     if (pt_phys == -1UL)
         panic("Can't initialize temp mapper");
