+++ kernel/generic/memory/mem_object.cc
@@ -522,4 +522,4 @@ ReturnStr<void *> Mem_Object::map_to_kernel(u64 offset, u64 size, Page_Table_Arg
     return mem_virt;
 }
 
-size_t Mem_Object::size_bytes() const noexcept { return pages_size << page_size_log; }
+u64 Mem_Object::size_bytes() const noexcept { return pages_size << page_size_log; }
